telemetry/coverage: pick the right clock per thread topology, and guard the cast
Follow-up to 6ef1c47, which fixed the pump but over-applied its fix and left the
sibling file half-hardened.

coverage.cpp: revert to emscripten_get_now(), widened to uint64_t instead of
truncated. The defect there was always the 32-bit cast, not the function. Unlike
the pump, this clock is not single-consumer: g_lastFlushMs is stamped by
flush_tally_locked(), which dx8wasm_get_coverage() reaches from any thread (see
the g_tallyLock note in the same file — a main-thread ccall really does run
concurrently with the engine worker under PROXY_TO_PTHREAD), while the
comparison runs on the producer thread. performance.now() has a per-thread
origin, so mixing bases in one cursor would underflow the unsigned subtract and
force a flush on the next token — not a correctness bug, it cannot fabricate or
lose counts and self-heals after one flush, but it defeats coalescing exactly
when a reader polls, adding the ring pressure the tally exists to avoid.
timeOrigin + performance.now() is the same base on every thread, which is what a
cross-thread cursor needs. The comment now states that, instead of inheriting
the pump's single-thread justification.

coverage.cpp: apply the g_everFlushed hardening the previous commit declared
necessary and then only did in telemetry.cpp. `g_lastFlushMs == 0` as "never
stamped" is wrong for the same reason there: 0 is a legal timestamp. Also stamp
the flag in flush_tally_locked(), or a get_coverage()-driven flush landing
before the first tally_record() would be discarded by that call's seeding
branch.

telemetry.cpp: the non-Emscripten path had regressed from "flush every call" to
"flush once, then never" — now_ms() returns a constant there, and the new
g_everFlushed flag made the gate latch on it. That is the very bug being fixed,
reintroduced on the portability path. Make that path's rate-limiter input
advance a full interval per call, and say why it must never be a constant.

telemetry.h: the pump's single-thread affinity is now load-bearing (a
thread-relative clock plus a non-atomic cursor), so state it as a requirement
rather than leaving "call it every frame" to imply it.

compatlib/win32_time.cpp: same defect, still live. timeGetTime() and
GetTickCount() cast the double straight to DWORD and so returned a constant
0xFFFFFFFF under -pthread — a frozen clock breaking every caller that paces or
times out. Widen to uint64_t first, then truncate, which gives the ~49.7-day
wrap the real Win32 API has and callers are written for.

scripts/check.sh: new guardrail 4 rejects any 32-bit cast of
emscripten_get_now() in tracked sources. This class of bug is invisible — no
warning, no trap, just a constant — and it has now appeared three times in this
repo. Verified the guardrail actually fails: restoring the old win32_time.cpp
lines makes check.sh report both and exit non-zero.

All 31 GPU smokes green (web-runtime/test/phase2.gpu.test.mjs, exit 0),
telemetry_smoke [3,1,1,1] and coverage_smoke [5,1,4,1] included.

+++ runtime/compatlib/win32_time.cpp
@@ -9,8 +9,17 @@
 
 extern "C" {
 
-DWORD timeGetTime(void) { return (DWORD)emscripten_get_now(); }
-DWORD GetTickCount(void) { return (DWORD)emscripten_get_now(); }
+// Both of these are 32-bit tick counts that the real Win32 API wraps every ~49.7
+// days, and callers are written for that. Getting there needs the double widened to
+// a 64-bit integer FIRST and only then truncated: casting the double straight to
+// DWORD is a float->i32 conversion, and under -pthread emscripten_get_now() is
+// performance.timeOrigin + performance.now() (~1.7e12 ms), far outside uint32_t's
+// range. wasm's non-trapping fptoui saturates rather than wrapping, so the direct
+// cast returned a constant 0xFFFFFFFF — a frozen clock, which silently breaks every
+// caller that paces or times out on it. Widen, then wrap. (Same defect as the one
+// fixed in telemetry.cpp/coverage.cpp; scripts/check.sh guards the pattern now.)
+DWORD timeGetTime(void) { return (DWORD)(uint64_t)emscripten_get_now(); }
+DWORD GetTickCount(void) { return (DWORD)(uint64_t)emscripten_get_now(); }
 
 BOOL QueryPerformanceFrequency(LARGE_INTEGER* freq) {
   if (!freq) return 0;
